pyproject.toml
[project]
name = "voice-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-google",
    "livekit-plugins-sarvam",
    "python-dotenv>=1.1.1",
]
